payment mechanism update
Payment mechanism for transaction fees. This feature is designed to ensure that users pay the transaction fee before executing a transaction. The fee structure is based on the number of previous transactions made by the user. The contract uses the transfer function to send the transaction fee from the user's account to the contract's account.

This payment mechanism is based on the assumption that users have enough funds in their account to cover the transaction fee. If a user does not have enough funds, the transfer function will return false, and the transaction will not execute.