payment mechanism update
This updated includes a payment mechanism for transaction fees using an ERC20 token. The deposit function includes a check to ensure that the user has enough tokens to make the deposit, by calling the balanceOf function from the ERC20 token contract. The contract also calls the approve function to allow the contract to spend certain amount of tokens on behalf of the user, and then the contract calls the transferFrom function to transfer the tokens from the user's account to the contract's account.

This improves interoperability by allowing the contract to work with any ERC20 token, as long as the token implements the ERC20 interface. This allows the contract to work with different tokens and makes it more versatile.